pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "ademirsoares",email = "ademir@licitaronline.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "sympy (>=1.14.0,<2.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=4.1,<5.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
